Refactor NeoFaker.App module
